Add demo communication between React App and Express App
To React app is properly communicating with the Express app, load the page (localhost:3000/). If the page reads "Welcome to React / API is working properly" then everything is good.

Note I have the React-Server relationship configured where React is running on port 3000, and Express on port 9000.

Changes to App.js were based off of the tutorial: https://www.freecodecamp.org/news/create-a-react-frontend-a-node-express-backend-and-connect-them-together-c5798926047c/